Break out tests and handle null node race
