+++ test/spec/picker/Picker.test.js
@@ -365,241 +365,4 @@ describe('Picker tests', () => {
       unicode: '🐒'
     }))
   })
-
-  test('press enter to make first search item active', async () => {
-    await tick(120)
-    type(getByRole('combobox'), 'monkey face')
-    await waitFor(() => expect(getAllByRole('option')).toHaveLength(1))
-    expect(getByRole('combobox').getAttribute('aria-activedescendant')).toBeFalsy()
-    await tick(120)
-    fireEvent.keyDown(getByRole('combobox'), { key: 'Enter', code: 'Enter' })
-    await waitFor(() => (
-      expect(getByRole('combobox').getAttribute('aria-activedescendant'))
-        .toBe(getByRole('option', { name: /🐵/ }).getAttribute('id'))
-    ), { timeout: 5000 })
-  }, 10000)
-
-  test('press enter on an empty search list', async () => {
-    await tick(120)
-    type(getByRole('combobox'), 'xxxyyyzzzhahaha')
-    await waitFor(() => expect(queryAllByRole('option')).toHaveLength(0))
-    expect(getByRole('combobox').getAttribute('aria-activedescendant')).toBeFalsy()
-    await tick(120)
-    fireEvent.keyDown(getByRole('combobox'), { key: 'Enter', code: 'Enter' })
-    await tick(120)
-    // should do nothing basically since there's nothing to search for
-    expect(queryAllByRole('option')).toHaveLength(0)
-    expect(getByRole('combobox').getAttribute('aria-activedescendant')).toBeFalsy()
-  }, 10000)
-
-  test('press enter to make first search item active - custom emoji', async () => {
-    picker.customEmoji = [
-      {
-        name: 'donkey',
-        shortcodes: ['donkey'],
-        url: 'donkey.png',
-        category: 'Ungulates'
-      }
-    ]
-
-    await tick(120)
-    type(getByRole('combobox'), 'donkey')
-    await waitFor(() => expect(getAllByRole('option')).toHaveLength(1))
-    expect(getByRole('combobox').getAttribute('aria-activedescendant')).toBeFalsy()
-    await tick(120)
-    fireEvent.keyDown(getByRole('combobox'), { key: 'Enter', code: 'Enter' })
-    await waitFor(() => {
-      return expect(getByRole('combobox').getAttribute('aria-activedescendant'))
-        .toBe(getByRole('option', { name: /donkey/ }).getAttribute('id'))
-    }, { timeout: 5000 })
-
-    let emoji
-    picker.addEventListener('emoji-click', event => {
-      emoji = event.detail
-    })
-
-    fireEvent.keyDown(getByRole('combobox'), { key: 'Enter', code: 'Enter' })
-
-    await waitFor(() => expect(emoji && emoji.name === 'donkey'))
-  }, 10000)
-
-  test('Closes skintone picker when blurred', async () => {
-    fireEvent.click(getByRole('button', { name: /Choose a skin tone/ }))
-    await waitFor(() => expect(getByRole('listbox', { name: 'Skin tones' })).toBeVisible())
-    // Simulating a focusout event is hard, have to both focus and blur
-    getByRole('combobox').focus()
-    fireEvent.focusOut(getByRole('listbox', { name: 'Skin tones' }))
-    await waitFor(() => expect(queryAllByRole('listbox', { name: 'Skin tones' })).toHaveLength(0))
-  })
-
-  test('Closes skintone picker when focus moves to skintone trigger button', async () => {
-    const chooseSkintoneButton = getByRole('button', { name: /Choose a skin tone/ })
-    fireEvent.click(chooseSkintoneButton)
-    await waitFor(() => expect(getByRole('listbox', { name: 'Skin tones' })).toBeVisible())
-    // Simulating a focusout event is hard, have to both focus and blur
-    chooseSkintoneButton.focus()
-    fireEvent.focusOut(getByRole('listbox', { name: 'Skin tones' }))
-    await waitFor(() => expect(queryAllByRole('listbox', { name: 'Skin tones' })).toHaveLength(0))
-  })
-
-  test('Custom emoji with categories', async () => {
-    picker.customEmoji = [
-      {
-        name: 'monkey',
-        shortcodes: ['monkey'],
-        url: 'monkey.png',
-        category: 'Primates'
-      },
-      {
-        name: 'donkey',
-        shortcodes: ['donkey'],
-        url: 'donkey.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'horse',
-        shortcodes: ['horse'],
-        url: 'horse.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'human',
-        shortcodes: ['human'],
-        url: 'human.png'
-      }
-    ]
-    await waitFor(() => expect(getAllByRole('tab')).toHaveLength(groups.length + 1))
-    await waitFor(() => expect(getAllByRole('menu')).toHaveLength(4)) // favorites + three custom categories
-    await waitFor(() => expect(getByRole('menuitem', { name: 'human' })).toBeVisible())
-    await waitFor(() => expect(getByRole('menuitem', { name: 'donkey' })).toBeVisible())
-    await waitFor(() => expect(getByRole('menuitem', { name: 'monkey' })).toBeVisible())
-    await waitFor(() => expect(getByRole('menuitem', { name: 'horse' })).toBeVisible())
-
-    // confirm alphabetical order for categories
-    expect(
-      await Promise.all(getAllByRole('menu').map(node => getAccessibleName(container, node)))
-    ).toStrictEqual(
-      ['Custom', 'Primates', 'Ungulates', 'Favorites']
-    )
-
-    // try searching
-    await type(getByRole('combobox'), 'donkey')
-    await waitFor(() => expect(getByRole('option', { name: 'donkey' })).toBeVisible())
-  })
-
-  test('Custom emoji with sorted categories and no shortcodes', async () => {
-    picker.customEmoji = [
-      {
-        name: 'monkey',
-        url: 'monkey.png',
-        category: 'Primates'
-      },
-      {
-        name: 'donkey',
-        url: 'donkey.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'horse',
-        url: 'horse.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'bird',
-        url: 'bird.png',
-        category: 'Avians'
-      },
-      {
-        name: 'human',
-        url: 'human.png'
-      }
-    ]
-    await waitFor(() => expect(getAllByRole('tab')).toHaveLength(groups.length + 1))
-    await waitFor(() => expect(getAllByRole('menu')).toHaveLength(5)) // favorites + four custom categories
-
-    // confirm alphabetical order for categories
-    expect(
-      await Promise.all(getAllByRole('menu').map(node => getAccessibleName(container, node)))
-    ).toStrictEqual([
-      'Custom',
-      'Avians',
-      'Primates',
-      'Ungulates',
-      'Favorites'
-    ])
-
-    const order = ['Ungulates', 'Primates', 'Avians']
-    picker.customCategorySorting = (a, b) => {
-      const aIdx = order.indexOf(a)
-      const bIdx = order.indexOf(b)
-      return aIdx < bIdx ? -1 : 1
-    }
-
-    await waitFor(async () => (
-      expect(
-        await Promise.all(getAllByRole('menu').map(node => getAccessibleName(container, node)))
-      ).toStrictEqual([
-        'Custom',
-        ...order,
-        'Favorites'
-      ])
-    ))
-  })
-
-  test('Custom emoji with all the same category', async () => {
-    picker.customEmoji = [
-      {
-        name: 'sheep',
-        url: 'sheep.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'deer',
-        url: 'deer.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'pig',
-        url: 'pig.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'horse',
-        url: 'horse.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'donkey',
-        url: 'donkey.png',
-        category: 'Ungulates'
-      },
-      {
-        name: 'rhinoceros',
-        url: 'rhinoceros.png',
-        category: 'Ungulates'
-      }
-    ]
-
-    await waitFor(() => expect(getAllByRole('tab')).toHaveLength(groups.length + 1))
-    await waitFor(() => expect(getAllByRole('menu')).toHaveLength(2)) // favorites + 1 custom categories
-
-    await waitFor(async () => (
-      expect(
-        await Promise.all(getAllByRole('menu').map(node => getAccessibleName(container, node)))
-      ).toStrictEqual([
-        'Ungulates',
-        'Favorites'
-      ])
-    ))
-
-    // Visibility test, has nothing to do with accessibility. We visually show the label if there's a single category
-    // and it's not the default "Custom" one.
-    expect(container.querySelector('.category').textContent).toEqual('Ungulates')
-    expect(container.querySelector('.category')).not.toHaveClass('gone')
-  })
-
-  test('Styles are working in Jest', async () => {
-    const style = picker.shadowRoot.querySelector('style')
-    expect(style.textContent).not.toEqual('')
-  })
 })
